+++ androidtoggleswitch-sample/src/main/java/com/example/androidtoggleswitch_sample/CustomSamplesFragment.java
@@ -8,6 +8,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
+import belka.us.androidtoggleswitch.widgets.MultipleToggleSwitch;
 import belka.us.androidtoggleswitch.widgets.ToggleSwitch;
 
 /**
@@ -20,57 +21,58 @@ public class CustomSamplesFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.custom_samples_fragment, container, false);
 
-//        final ToggleSwitch matchParentWidthToggleSwitch = (ToggleSwitch) view.findViewById(R.id.matchParentWidthToggleSwitch);
-//        final ToggleSwitch customColorsToggleSwitch = (ToggleSwitch) view.findViewById(R.id.customColorsToggleSwitch);
-//        final ToggleSwitch customSizesToggleSwitch = (ToggleSwitch) view.findViewById(R.id.customSizesToggleSwitch);
-//        final MultipleToggleSwitch customBordersToggleSwitch = (MultipleToggleSwitch) view.findViewById(R.id.customBordersToggleSwitch);
-//        final ToggleSwitch noSeparatorToggleSwitch = (ToggleSwitch) view.findViewById(R.id.noSeparatorToggleSwitch);
-//
-//
+        final ToggleSwitch matchParentWidthToggleSwitch = (ToggleSwitch) view.findViewById(R.id.matchParentWidthToggleSwitch);
+        final ToggleSwitch customColorsToggleSwitch = (ToggleSwitch) view.findViewById(R.id.customColorsToggleSwitch);
+        final ToggleSwitch customSizesToggleSwitch = (ToggleSwitch) view.findViewById(R.id.customSizesToggleSwitch);
+        final MultipleToggleSwitch customBordersToggleSwitch = (MultipleToggleSwitch) view.findViewById(R.id.customBordersToggleSwitch);
+        final ToggleSwitch noSeparatorToggleSwitch = (ToggleSwitch) view.findViewById(R.id.noSeparatorToggleSwitch);
+
+
 //        matchParentWidthToggleSwitch.setOnChangeListener(new ToggleSwitch.OnChangeListener() {
 //            @Override
-//            public void onToggleSwitchChanged(int position) {
-//                showClickedToast(matchParentWidthToggleSwitch, position);
+//            public void onToggleSwitchChanged(int pos) {
+//                showClickedToast(matchParentWidthToggleSwitch, pos);
 //            }
 //        });
 //
 //        customColorsToggleSwitch.setOnChangeListener(new ToggleSwitch.OnChangeListener() {
 //            @Override
-//            public void onToggleSwitchChanged(int position) {
-//                showClickedToast(customColorsToggleSwitch, position);
+//            public void onToggleSwitchChanged(int pos) {
+//                showClickedToast(customColorsToggleSwitch, pos);
 //            }
 //        });
 //
 //        customSizesToggleSwitch.setOnChangeListener(new ToggleSwitch.OnChangeListener() {
 //            @Override
-//            public void onToggleSwitchChanged(int position) {
-//                showClickedToast(customSizesToggleSwitch, position);
+//            public void onToggleSwitchChanged(int pos) {
+//                showClickedToast(customSizesToggleSwitch, pos);
 //            }
 //        });
 //
 //        noSeparatorToggleSwitch.setOnChangeListener(new ToggleSwitch.OnChangeListener() {
 //            @Override
-//            public void onToggleSwitchChanged(int position) {
-//                showClickedToast(noSeparatorToggleSwitch, position);
+//            public void onToggleSwitchChanged(int pos) {
+//                showClickedToast(noSeparatorToggleSwitch, pos);
 //            }
 //        });
 
-//        customBordersToggleSwitch.setOnChangeListener(new MultipleToggleSwitch.OnChangeListener() {
-//            @Override
-//            public void onMultipleToggleSwitchChanged(int position, boolean checked) {
-//                String label = customBordersToggleSwitch.getLabels().get(position);
-//                Toast.makeText(getActivity(),
-//                        label + "[" + position + "] => " +  checked,
-//                        Toast.LENGTH_SHORT).show();
-//            }
-//        });
+        customBordersToggleSwitch.setOnChangeListener(new MultipleToggleSwitch.OnChangeListener() {
+            @Override
+            public void onMultipleToggleSwitchChanged(int position, boolean checked) {
+                String[] planets = getResources().getStringArray(R.array.planets);
+                String label = planets[position];
+                Toast.makeText(getActivity(),
+                        label + "[" + position + "] => " +  checked,
+                        Toast.LENGTH_SHORT).show();
+            }
+        });
 
         return view;
     }
 
-    private void showClickedToast(ToggleSwitch toggleSwitch, int position) {
-        Toast.makeText(getActivity(),
-                "Checked: " + toggleSwitch.getLabels().get(position),
-                Toast.LENGTH_SHORT).show();
-    }
+//    private void showClickedToast(ToggleSwitch toggleSwitch, int pos) {
+//        Toast.makeText(getActivity(),
+//                "Checked: " + toggleSwitch.getLabels().get(pos),
+//                Toast.LENGTH_SHORT).show();
+//    }
 }