+++ androidtoggleswitch/src/main/java/belka/us/androidtoggleswitch/widgets/IRightToLeftProvider.kt
@@ -1,0 +1,9 @@
+package belka.us.androidtoggleswitch.widgets
+
+/**
+ * Created by lorenzorigato on 24/06/2017.
+ */
+
+interface IRightToLeftProvider {
+    fun isRightToLeft(): Boolean
+}